WIP - add suspense boundary
