+++ apps/web/src/app/auth/desktop/success/components/DesktopAuthSuccess/index.ts
@@ -1,0 +1,1 @@
+export { DesktopAuthSuccess } from "./DesktopAuthSuccess";
