+++ apps/web/src/app/auth/desktop/success/components/LoadingFallback/index.ts
@@ -1,0 +1,1 @@
+export { LoadingFallback } from "./LoadingFallback";
